Convert filter builtin to class
